fix: <a> tag in footer to open in a new tab
If the content of the sidebar is long enough the users might misclick
and choose the *Edit this page* link instead of the *Next: Lesson* link.
By doing so, the link will open within the user lesson tab and even if
they go back to their previous tab the progress will be lost.

This patch makes the <a> tag on the footer open in a new tab, so if
the user misclicks on the link they can close it and continue normally.