fix(02-advanced-svelte/07-composition): apply missing fill to `paper.svg`
When used as a background image, not including a fill will make the svg
invisible, this commit adds the fill so the example works properly.

Squashed the following changes in as well because the asset was also
broken in subsequent tutorials. I have aligned them with the changes I
made in `07-composition/01-slots`.